vet_clinic_django/backend/app/requirements.txt
asgiref==3.8.1
coloredlogs==15.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
humanfriendly==10.0
Markdown==3.7
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
